Define test_config, used by both jest and test server
