+++ tests/integration/server_start.js
@@ -1,6 +1,6 @@
 const App = require( './../../bin/app');
-const configBuilder = require('@qwant/nconf-builder');
 const nock = require('nock');
+const config = require('./test_config');
 
 const { ...poiNoName } = require('../__data__/poi.json');
 /* default test with matching name & local_name */
@@ -29,20 +29,6 @@ nock(/idunn_test\.test/)
   .get(/osm:way:2403/)
   .reply(404);
 
-const config = configBuilder.get_without_check();
-
-// Specific config values for tests
-config.mapStyle.baseMapUrl = '[]';
-config.mapStyle.poiMapUrl = '[]';
-config.services.idunn.url = 'http://idunn_test.test';
-config.services.geocoder.url = 'http://geocoder.test/autocomplete';
-config.direction.enabled = true;
-config.direction.service.api = 'mapbox'; // Directions fixtures use mapbox format
-config.category.enabled = true;
-config.events.enabled = true;
-config.masq.enabled = false;
-config.system.baseUrl = '/maps/';
-config.server.routerBaseUrl = '/maps/';
 
 global.appServer = new App(config);
